% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{histNdistri}
\alias{histNdistri}
\title{plot KTU evaluations}
\usage{
histNdistri(
  histdata,
  coll,
  xlab.text = NULL,
  legend.posit = "topright",
  legend.text = "",
  ...
)
}
\arguments{
\item{histdata}{KTU evaluation variable}

\item{coll}{color for distribution curve}
}
\description{
Evaluate sequence similarity within KTUs by histogram
}
