% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{kaxonomy}
\alias{kaxonomy}
\title{KTU annotation}
\usage{
kaxonomy(
  dbRDS,
  taxaRDS,
  kmer.table,
  cos.cutff = 0.95,
  consensus = 0.8,
  candidate = 10,
  cores = 1
)
}
\arguments{
\item{dbRDS}{Kmer formated database RDS file}

\item{taxaRDS}{taxonomy ID RDS file}

\item{kmer.table}{tetranucleotide frequency table for annotation}

\item{cos.cutff}{cosine similarity cutoff, default=0.95}

\item{consensus}{taxonomy assignment by consensus [0-1]}

\item{candidate}{how many candidates (≥ cos.cutff) for taxonomy assignment}

\item{cores}{Numbers of CPUs}
}
\description{
KTU-based alignment-free taxonomy assignment
}
