% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{KTUsim.eval}
\alias{KTUsim.eval}
\title{KTU evaluation}
\usage{
KTUsim.eval(klusterRDS, ASVfasta)
}
\arguments{
\item{klusterRDS}{klustering output RDS file}

\item{ASVfasta}{representative ASV sequence fasta file}
}
\description{
Evaluate sequence similarity within KTUs
}
