% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{makektudb}
\alias{makektudb}
\title{Build KTU taxonomy assignment database by tetranucleotide frequency table}
\usage{
makektudb(input.fasta, input.taxa, output.file = NULL)
}
\arguments{
\item{input.fasta}{input database fasta file (trimmed)}

\item{input.taxa}{input database taxonomy file}

\item{output.file}{(optional) the directory for output RDS format file}
}
\description{
make KTU db
}
