% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{trim.primer}
\alias{trim.primer}
\title{Prepare KTU taxonomy assignment database by trimming primers}
\usage{
trim.primer(fasta, forseq, revseq, output.name = NULL, progress = TRUE)
}
\arguments{
\item{fasta}{input database fasta file}

\item{forseq}{forward primer sequence (5' -> 3')}

\item{revseq}{reverse primer sequence (5' -> 3')}

\item{output.name}{(optional) the name for output file}

\item{progress}{show the processing progress}
}
\description{
Trim primer sequences of full length 16S/18S/target gene database by PCR primers
}
