% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FN_KTU.R
\name{klustering}
\alias{klustering}
\title{KTU clustering}
\usage{
klustering(
  repseq,
  feature.table = NULL,
  write.fasta = TRUE,
  step = c(3, 10),
  search.min = 10,
  search.max = NULL,
  cores = 1
)
}
\arguments{
\item{repseq}{The fasta file path}

\item{feature.table}{(optional) 'data.frame' formated ASV/OTU table}

\item{write.fasta}{(optional) write out a representative KTU sequence fasta file}

\item{step}{Split searching range for optimal K. Two step searching process: large scale searching in the first round and smaller scale searching in the second round. Default is 'step=c(3,10)'.}

\item{search.min}{The minimum K for searching, default 'search.min=10'}

\item{search.max}{The maximum K for searching, default is 'NULL' = max}

\item{cores}{Numbers of CPUs}
}
\value{
KTU.table aggregated KTU table

ReqSeq Representative KTU sequences

kmer.table Tetranucleotide frequency table

clusters K-clusters of input features
}
\description{
K-mer taxonomic unit clustering
}
